requirements.txt
bitstring==4.0.1
numpy==1.24.2
Pillow==9.4.0
